Revert "Revert all changes commited on Sep 20 by @jeanschmidt (#758)"
This reverts commit da1204abc9ad4962616ae5852127198ac695eb8f.

Need it to run terraform plan
