+++ terraform-aws-github-runner/main.tf
@@ -36,7 +36,6 @@ resource "aws_sqs_queue" "queued_builds" {
 module "webhook" {
   source = "./modules/webhook"
 
-  aws_region  = var.aws_region
   environment = var.environment
   tags        = local.tags
   encryption = {
@@ -81,7 +80,6 @@ module "runners" {
   cant_have_issues_labels = var.cant_have_issues_labels
 
   instance_type         = var.instance_type
-  market_options        = var.market_options
   block_device_mappings = var.block_device_mappings
 
   runner_architecture = local.runner_architecture
@@ -96,7 +94,6 @@ module "runners" {
   minimum_running_time_in_minutes      = var.minimum_running_time_in_minutes
   runner_extra_labels                  = var.runner_extra_labels
   runner_as_root                       = var.runner_as_root
-  runners_maximum_count                = var.runners_maximum_count
   idle_config                          = var.idle_config
   enable_ssm_on_runners                = var.enable_ssm_on_runners
   runner_additional_security_group_ids = var.runner_additional_security_group_ids
@@ -113,8 +110,6 @@ module "runners" {
   logging_retention_in_days        = var.logging_retention_in_days
   enable_cloudwatch_agent          = var.enable_cloudwatch_agent
   cloudwatch_config                = var.cloudwatch_config
-  runner_log_files                 = var.runner_log_files
-  runner_group_name                = var.runner_group_name
   scale_up_lambda_concurrency      = var.scale_up_lambda_concurrency
 
   instance_profile_path     = var.instance_profile_path
@@ -136,17 +131,14 @@ module "runners" {
 module "runner_binaries" {
   source = "./modules/runner-binaries-syncer"
 
-  aws_region  = var.aws_region
   environment = var.environment
   tags        = local.tags
 
   distribution_bucket_name = "${var.environment}-dist-${random_string.random.result}"
 
-  runner_architecture              = local.runner_architecture
   runner_allow_prerelease_binaries = var.runner_allow_prerelease_binaries
 
   lambda_s3_bucket                = var.lambda_s3_bucket
-  syncer_lambda_s3_key            = var.syncer_lambda_s3_key
   syncer_lambda_s3_object_version = var.syncer_lambda_s3_object_version
   lambda_zip                      = var.runner_binaries_syncer_lambda_zip
   lambda_timeout                  = var.runner_binaries_syncer_lambda_timeout
