+++ terraform-aws-github-runner/modules/runners/lambdas/runners/src/scale-runners/gh-issues.test.ts
@@ -1,12 +1,13 @@
 import { getRepoIssuesWithLabel, resetIssuesCache } from './gh-issues';
 
 import { Octokit } from '@octokit/rest';
-import { createGitHubClientForRunnerRepo } from './runners';
+import { createGitHubClientForRunnerRepo } from './gh-runners';
 import { mocked } from 'ts-jest/utils';
 import nock from 'nock';
 import { ScaleUpMetrics } from './metrics';
 
 jest.mock('./runners');
+jest.mock('./gh-runners');
 
 beforeEach(() => {
   jest.resetModules();
