+++ terraform-aws-github-runner/modules/runners/lambdas/runners/src/scale-runners/scale-down.test.ts
@@ -1,15 +1,13 @@
+import { listRunners, resetRunnersCaches, terminateRunner } from './runners';
 import {
   GhRunners,
   getRunnerOrg,
   getRunnerRepo,
   listGithubRunnersOrg,
   listGithubRunnersRepo,
-  listRunners,
   removeGithubRunnerOrg,
   removeGithubRunnerRepo,
-  resetRunnersCaches,
-  terminateRunner,
-} from './runners';
+} from './gh-runners';
 
 import { Config } from './config';
 import { mocked } from 'ts-jest/utils';
@@ -18,16 +16,23 @@ import nock from 'nock';
 import scaleDown from './scale-down';
 import * as MetricsModule from './metrics';
 
-jest.mock('./runners', () => ({
+jest.mock('./gh-runners', () => ({
   /* eslint-disable @typescript-eslint/no-explicit-any */
   ...(jest.requireActual('./runners') as any),
   getRunnerRepo: jest.fn(),
   getRunnerOrg: jest.fn(),
   listGithubRunnersRepo: jest.fn(),
   listGithubRunnersOrg: jest.fn(),
-  listRunners: jest.fn(),
   removeGithubRunnerOrg: jest.fn(),
   removeGithubRunnerRepo: jest.fn(),
+  resetGHRunnersCaches: jest.fn(),
+  terminateRunner: jest.fn(),
+}));
+
+jest.mock('./runners', () => ({
+  /* eslint-disable @typescript-eslint/no-explicit-any */
+  ...(jest.requireActual('./runners') as any),
+  listRunners: jest.fn(),
   resetRunnersCaches: jest.fn(),
   terminateRunner: jest.fn(),
 }));
