+++ terraform-aws-github-runner/modules/runners/lambdas/runners/src/scale-runners/scale-up.test.ts
@@ -1,11 +1,11 @@
+import { createRunner } from './runners';
 import {
   createRegistrationTokenOrg,
   createRegistrationTokenRepo,
-  createRunner,
   getRunnerTypes,
   listGithubRunnersOrg,
   listGithubRunnersRepo,
-} from './runners';
+} from './gh-runners';
 
 import { Config } from './config';
 import { getRepoIssuesWithLabel, GhIssues } from './gh-issues';
@@ -15,6 +15,7 @@ import { scaleUp } from './scale-up';
 import * as MetricsModule from './metrics';
 
 jest.mock('./runners');
+jest.mock('./gh-runners');
 jest.mock('./gh-issues');
 jest.mock('./metrics');
