+++ terraform-aws-github-runner/modules/runners/lambdas/runners/src/scale-runners/utils.ts
@@ -60,7 +60,7 @@ export async function expBackOff<T>(
     try {
       return await callback();
     } catch (e) {
-      if (`${e}`.includes('RequestLimitExceeded')) {
+      if (`${e}`.includes('RequestLimitExceeded') || `${e}`.includes('ThrottlingException')) {
         if (expBackOffMs > maxMs) {
           throw e;
         }
