+++ terraform-aws-github-runner/modules/runners/policies/instance-ssm-parameters-policy.json
@@ -3,6 +3,11 @@
   "Statement": [
     {
       "Effect": "Allow",
+      "Action": ["ssm:DescribeParameters"],
+      "Resource": ["*"]
+    },
+    {
+      "Effect": "Allow",
       "Action": ["ssm:DeleteParameter"],
       "Resource": "${arn_ssm_parameters}"
     },
