+++ terraform-aws-github-runner/modules/runners/policies/lambda-scale-down.json
@@ -19,6 +19,13 @@
     {
       "Effect": "Allow",
       "Action": [
+        "ssm:DescribeParameters"
+      ],
+      "Resource": ["*"]
+    },
+    {
+      "Effect": "Allow",
+      "Action": [
         "ec2:TerminateInstances"
       ],
       "Resource": ["*"],
