+++ terraform-aws-github-runner/modules/runners/scale-down.tf
@@ -28,6 +28,7 @@ resource "aws_lambda_function" "scale_down" {
 
   environment {
     variables = {
+      AWS_REGION_INSTANCES            = join(",", var.aws_region_instances)
       ENABLE_ORGANIZATION_RUNNERS     = var.enable_organization_runners
       ENVIRONMENT                     = var.environment
       GHES_URL                        = var.ghes_url
